+++ MangArchipelBack/.settings/org.eclipse.jdt.core.prefs
