+++ MangArchipelBack/src/main/java/MangArchipelBack/App.java
@@ -1,0 +1,13 @@
+package MangArchipelBack;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
