+++ MangArchipelFront/.settings/org.eclipse.jdt.core.prefs
