+++ MangArchipelFront/src/main/java/MangArchipelFront/App.java
@@ -1,0 +1,13 @@
+package MangArchipelFront;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
