+++ pom.xml
@@ -1,25 +1,134 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
 
-  <groupId>fr.dtaingenierie</groupId>
-  <artifactId>MangArchipel</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
+	<groupId>fr.dtaingenierie</groupId>
+	<artifactId>MangArchipel</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<packaging>pom</packaging>
 
-  <name>MangArchipel</name>
-  <url>http://maven.apache.org</url>
+	<name>MangArchipel</name>
+	<url>http://maven.apache.org</url>
+	
+	<developers>
+		<developer>
+			<id>xhackax47</id>
+			<name>Samy</name>
+			<email>xhackax47@gmail.com</email>
+			<url>https://github.com/xhackax47</url>
+			<organization>NONE</organization>
+			<organizationUrl/>
+			<roles>
+				<role>Architect</role>
+				<role>Java /J2EE Developer</role>
+			</roles>
+			<timezone>France/Montpellier</timezone>
+			<properties>
+				<picUrl/>
+			</properties>
+		</developer>
 
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
+		<developer>
+			<id/>
+			<name>Bruno</name>
+			<email/>
+			<url/>
+			<organization>DTA INGENIERIE</organization>
+			<organizationUrl/>
+			<roles>
+				<role>Architect</role>
+				<role>Java /J2EE Developer</role>
+			</roles>
+			<timezone>France/Montpellier</timezone>
+			<properties>
+				<picUrl/>
+			</properties>
+		</developer>
 
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
+		<developer>
+			<id/>
+			<name>Angelina</name>
+			<email/>
+			<url/>
+			<organization>DTA INGENIERIE</organization>
+			<organizationUrl/>
+			<roles>
+				<role>Architect</role>
+				<role>Java /J2EE Developer</role>
+			</roles>
+			<timezone>France/Montpellier</timezone>
+			<properties>
+				<picUrl/>
+			</properties>
+		</developer>
+
+		<developer>
+			<id/>
+			<name>Florent</name>
+			<email/>
+			<url/>
+			<organization>DTA INGENIERIE</organization>
+			<organizationUrl/>
+			<roles>
+				<role>Architect</role>
+				<role>Java /J2EE Developer</role>
+			</roles>
+			<timezone>France/Montpellier</timezone>
+			<properties>
+				<picUrl/>
+			</properties>
+		</developer>
+	</developers>
+
+	<properties>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+		<project.build.outputEncoding>UTF-8</project.build.outputEncoding>
+		<project.build.inputEncoding>UTF-8</project.build.inputEncoding>
+		<maven.compiler.target>1.8</maven.compiler.target>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.inputEncoding>UTF-8</project.reporting.inputEncoding>
+	</properties>
+
+	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.0</version>
+				<executions>
+					<execution>
+						<id>default-compile</id>
+						<phase>compile</phase>
+						<goals>
+							<goal>compile</goal>
+						</goals>
+						<configuration>
+							<source>1.8</source>
+							<target>1.8</target>
+						</configuration>
+					</execution>
+					<execution>
+						<id>default-testCompile</id>
+						<phase>test-compile</phase>
+						<goals>
+							<goal>testCompile</goal>
+						</goals>
+						<configuration>
+							<source>1.8</source>
+							<target>1.8</target>
+						</configuration>
+					</execution>
+				</executions>
+				<configuration>
+					<source>1.8</source>
+					<target>1.8</target>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+
+	<modules>
+		<module>MangArchipelBack</module>
+		<module>MangArchipelFront</module>
+	</modules>
 </project>
