--- src/main/java/fr/dtaingenierie/MangArchipel/App.java
@@ -1,13 +1,0 @@
-package fr.dtaingenierie.MangArchipel;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
